Conversion to bootstrap instead of semantic UI
The angular components for Bootstrap are in a much better state, so this
helps with the development. Made it possible to remove the pagination
module, and also the date selectors work in Safari and Firefox now.

This should also make it easier for people to apply custom themes.

Removes dependency on jquery completely.